fix(app): avoid get foods when current search term is the same as previous term
- Added validation when submit the form to only set states with searchable parameters
when the search term is new

After adding the validation aforementioned, the search by already filtered results (aka cached results)
stopped working, because the `useEffect` that updates the `cachedFoodsResults` was not running the setState,
due to the check `hasRecentlyFetched`. Since the result was already cached by apollo, the
loading didn't change. To bridle this behaviour, a ref was added when submitting,
and updated the `useEffect` to user the new ref as comparison method.